chore(cli): refactor operators command in files
